rename components due to conflicts
