+++ addon/templates/components/adde-color-picker-cell.hbs
