+++ addon/templates/components/adde-color-picker-radio-cell.hbs
@@ -3,7 +3,7 @@
   value={{color}}
   checked={{isSelected}}
   data-close>
-{{color-picker-cell
+{{adde-color-picker-cell
   color=color
   data-test-color-cell=true
 }}
