+++ addon/templates/components/adde-color-picker.hbs
@@ -5,7 +5,7 @@
   {{#if _isEmpty}}
     &nbsp; + Color
   {{else}}
-    {{color-picker-cell color=selectedColor data-test-color-cell=true}}
+    {{adde-color-picker-cell color=selectedColor data-test-color-cell=true}}
   {{/if}}
   <span class="button-caret">
     <i class="fa fa-caret-down"></i>
@@ -21,7 +21,7 @@
         {{#each _colorPaletteAdjusted as |row|}}
           <div class="adde-color-picker-row">
             {{#each row as |color|}}
-              {{color-picker-radio-cell
+              {{adde-color-picker-radio-cell
                 color=color
                 selectedColor=selectedColor
                 data-test-radio-cell=true
@@ -36,7 +36,7 @@
           disabled={{_isCustomColorInvalid}}
           data-close
           data-test-custom-color-button>
-          {{color-picker-cell
+          {{adde-color-picker-cell
             color=_customColor
             class=(if _isCustomColorInvalid 'is-invalid' 'is-active')
             data-test-color-cell=true
