+++ app/components/adde-color-picker-radio-cell.js
@@ -1,0 +1,1 @@
+export { default } from '@addepar/forms/components/adde-color-picker-radio-cell';
