--- app/components/color-picker-radio-cell.js
@@ -1,1 +1,0 @@
-export { default } from '@addepar/forms/components/color-picker-radio-cell';
