+++ tests/integration/components/adde-color-picker-cell-test.js
@@ -5,14 +5,14 @@ import AddeColorPickerCellPage from '@addepar/forms/test-support/pages/adde-colo
 
 const ColorPickerCellHelper = AddeColorPickerCellPage.scope('[data-test-color-cell]');
 
-moduleForComponent('color-picker-cell', 'Integration | Component | color picker cell', {
+moduleForComponent('adde-color-picker-cell', 'Integration | Component | color picker cell', {
   integration: true,
 });
 
 test('Color cell renders with background color', function(assert) {
   assert.expect(1);
 
-  this.render(hbs`{{color-picker-cell color="#000000" data-test-color-cell=true}}`);
+  this.render(hbs`{{adde-color-picker-cell color="#000000" data-test-color-cell=true}}`);
 
   let cell = ColorPickerCellHelper.create();
 
@@ -22,7 +22,7 @@ test('Color cell renders with background color', function(assert) {
 test('Color cell correctly identifies as transparent', function(assert) {
   assert.expect(1);
 
-  this.render(hbs`{{color-picker-cell color="transparent" data-test-color-cell=true}}`);
+  this.render(hbs`{{adde-color-picker-cell color="transparent" data-test-color-cell=true}}`);
 
   let cell = ColorPickerCellHelper.create();
 
@@ -32,7 +32,7 @@ test('Color cell correctly identifies as transparent', function(assert) {
 test('Color cell correctly identifies as empty', function(assert) {
   assert.expect(1);
 
-  this.render(hbs`{{color-picker-cell color="" data-test-color-cell=true}}`);
+  this.render(hbs`{{adde-color-picker-cell color="" data-test-color-cell=true}}`);
 
   let cell = ColorPickerCellHelper.create();
